add test to assert num queries scales as expected
also move other test to dedicated file
